requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
mysqlclient==2.2.7
pillow==11.3.0
PyJWT==2.9.0
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
openpyxl==3.1.5
